Unneeded cast to float.

+++ src/rendering/render_box.cpp
@@ -1153,7 +1153,7 @@ void RenderBox::relativePositionOffset(int &tx, int &ty) const
     }
     if (!style()->top().isAuto()) {
         if (style()->top().isPercent()) {
-            float p = style()->top().percent();
+            double p = style()->top().percent();
             bool neg = p < 0.0;
             int ph = calcPercentageHeight(Length((neg ? -p : p), Percent));
             if (ph != -1) {
@@ -1164,7 +1164,7 @@ void RenderBox::relativePositionOffset(int &tx, int &ty) const
         }
     } else if (!style()->bottom().isAuto()) {
         if (style()->bottom().isPercent()) {
-            float p = style()->bottom().percent();
+            double p = style()->bottom().percent();
             bool neg = p < 0.0;
             int ph = calcPercentageHeight(Length((neg ? -p : p), Percent));
             if (ph != -1) {
